Use lower case on the SQLAnywhereJDBCAdapter
If the case is correct, we get multiple records inserted in the same
statement (which isn't supported in our version of SQLAnywhere).

If the adapter is not defined, we get an error saying the method import
can't be found for our model.

So I've switched to the middle ground which just works (define our
adapter but define it wrong). This way when we use import, we get
multiple insert statements.